RDKOSS-308: Fix sstate signinfo file download errors due to excessive filename length
Reason for the change:
In the current design, the maximum length for an sstate ".tar.zst" file is 246 characters, while the signinfo file can be up to 254 characters long. When these files are hosted on HTTP servers and accessed via SSTATE_MIRRORS, the signinfo files are temporarily downloaded using wget with a .tmp extension, resulting in <siginfo_file>.tmp. Once the download is complete, the file is renamed to remove the ".tmp" extension.

However, if the signinfo file's name is already 254 characters long, appending the .tmp extension increases the length to 258 characters, which exceeds the 256-character limit set by the Linux filesystem. Consequently, this results in a "file name too long" error during the download process.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>