yocto-bsp/6.16: specify genericarm64 SRCREV
We had previously dropped the genericarm64 as it was using
standard/base for builds. But we have some temporary serial
patches only for these platforms, so we we are building out
of standard/genericarm64 to isolate those patches.

As a result, we do need to set our SRCREV to ensure that
the content is built on the branch.

Note: we can also move these to the main linux-yocto recipes
as genericarm64 has broad scope and is more than just an
indidual hardware reference BSP.

(From meta-yocto rev: c20111924dae9106d439fe58d132cd64ae04478b)

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>