RDKOSS-480: Enable CVE check for packagegroup target
Reason for the change: By default, cve_check.bbclass enables CVE report generation only when an image or a specific recipe is explicitly built as a target. The fix extends that behavior to support CVE checks for packagegroup targets. It is particularly useful in RDKE-based stacks, where releases are typically driven by packagegroup recipes.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>