RDKOSS-480: Enable CVE check for packagegroup target (#23)
* RDKOSS-480: Enable CVE check for packagegroup target

Reason for the change: By default, cve_check.bbclass enables CVE report generation only when an image or a specific recipe is explicitly built as a target. The fix extends that behavior to support CVE checks for packagegroup targets. It is particularly useful in RDKE-based stacks, where releases are typically driven by packagegroup recipes.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>

* RDKOSS-480: Exclude prebuilt packages from cve report

Reason for the change: Packages without metadata in the current build workspace are assumed to be installed from the prebuilt IPK feed. Therefore, these packages are excluded from the target CVE report to fix the build failure.

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>

---------

Signed-off-by: mselva006c <mani_selvaraj@comcast.com>
Co-authored-by: mselva006c <mani_selvaraj@comcast.com>